Implemented the load instruction, but not its test
Implemented the load instruction. I found an issue with the TS type checker, a potential bug that we could patch ourselves.
I was not able to implement the test because i forgot to implement Memory.set(). I am going to implement that next, with its
test and then implement the test for the load instruction. Technically speacking I could do it without Memory.set() but I am
looking into implementing conditional testing.