Attempted to implement conditional execution
Partially implemented conditional execution of tests. The idea is that if a test depends on other being
successful, then the test will not run unless that test passes. The idea is that this can be tracked.
But after looking at the complexity of the code and how easy it is to break, I think it is better never
to try this ever again. It is just easier and better to make all test completely independent, even if
that means one needs to access the guts of a class and make the testing itself supper brittle and tedious.