Implemented some arithmetic tests in a generic style and fixed one nasty bug
It took me some time to finally realize that there was no way to decouple the set and get methods.
So I tested them both in the same test. The test will not reveal which broke, but this way the
implementation details are hidden away. I also found a bug that did not allow setting or getting
register $zero. Turns out that !0 (logic-not NumberZero) is evaluated as true by JS. I had to
change all the header guards. There might still be some instances of this bug so watch out. I intend
to give it another scan tomorrow. I used the find fn from the editor to find all the instances of '!'.
Since some of these are used appropriately, I could have missed one of the bugs. So watch out for this
mistake I made.