Implemented the store word instruction and its test
Implemented the store works instruction, and its test.
Fixed a nonfatal bug in  the initialization of the simulator. I forgot to set the
text and stack root addresses. Even though this is not fatal, it is not supposed
to happen. A check must be made that ensures that some sort of file is loaded
into memory or that a simulation object cannot exist until a file is provided.

In that case a file listener must be made, or a checksum of the file to be runned
made. In the even where the checksum changes, then we make a new simulator obj.
I prefer that one. It is simpler. All of that code is executed when the user
clicks the run button.